// src/interfaces/ITitle.ts
export default interface ITitle {
    id: string;
    title_id: string;
    owner_id: string;
    titulo: string;
    nota: string;
    stream: string;
    classificacao: string;
    status: string;
    tipo: string;
    image: string;
    background: string;
    sinopse: string;
    estrelas: any;
    genero: string;
};